apps/api/pyproject.toml
[project]
name = "openbb-platform-api"
version = "0.1.0"
description = "OpenBB personal information platform API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.4.0",
  "apscheduler>=3.10.4",
  "pyyaml>=6.0.2",
  "feedparser>=6.0.11",
  "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.2",
  "ruff>=0.7.0",
  "mypy>=1.12.0",
  "types-python-dateutil>=2.9.0.20241003",
  "types-PyYAML>=6.0.12.20241230",
]
market = ["openbb>=4.0.0"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --maxfail=1"

[tool.ruff]
line-length = 130
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
pretty = true
show_error_codes = true
ignore_missing_imports = true
